added some more option menu items, added BaseActivity class
