+++ SightLocator/res/values/strings.xml
@@ -3,6 +3,8 @@
 
     <string name="app_name">SightLocator</string>
     <string name="action_settings">Settings</string>
+    <string name="action_help">Help</string>
+    <string name="action_about">About</string>
     <string name="welcome_screen_text">The mobile development community is at a tipping point. Mobile users demand more choice, more opportunities to customize their phones, and more functionality. Mobile operators want to provide value-added content to their subscribers in a manageable and lucrative way. Mobile developers want the freedom to develop the powerful mobile applications users demand with minimal roadblocks to success. Finally, handset manufacturers want a stable, secure, and affordable platform to power their devices. Upuntil now single mobile platform has adequately addressed the needs of all the parties. Enter Android, which is a potential game-changer for the mobile development community. An innovative and open platform, Android is well positioned to address the growing needs of the mobile marketplace.</string>
     <string name="sts_olha_and_elisabeth">The Church of Sts. Olha and Elizabeth in Lviv, Ukraine, is located between the citys main rail station and the Old Town. It was originally built as the Roman Catholic Church of St. Elizabeth and today serves as the Greek Catholic Church of Sts. Olha and Elizabeth. The church was built by the Latin Archbishop of Lviv Saint Joseph Bilczewski in the years 1903&#8211;1911 as a parish church for the dynamically developing western suburb. It was designed by Polish architect Teodor Talowski</string>
     <string name="railway_station_wiki">The Main railway station in Lviv, Ukraine also known as Lviv-Main. It is one of the most notable pieces of Art Nouveau architecture in former Galicia. The station was opened to the public in 1904, and celebrated its centennial anniversary on 26 March 2004. On a monthly basis, the terminal handles over 1.2 million passengers and moves 16 thousand tons of freight.</string>
@@ -11,5 +13,5 @@
     <string name="imageview_tag_uri">0</string>
     <string name="display_full_screen_text">Info</string>
     <string name="address">Address : to do</string>
-    
+
 </resources>
