+++ SightLocator/src/com/iolab/sightlocator/MainActivity.java
@@ -1,19 +1,17 @@
 package com.iolab.sightlocator;
 
-import android.app.Activity;
-import android.app.Dialog;
-import android.os.Bundle;
-import android.view.Menu;
-
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GooglePlayServicesUtil;
 
-public class MainActivity extends Activity {
+import android.app.Dialog;
+import android.os.Bundle;
+
+public class MainActivity extends BaseActivity {
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-
+		
 		// checking if GooglePlayServices is installed
 		int status = GooglePlayServicesUtil
 				.isGooglePlayServicesAvailable(getBaseContext());
@@ -30,24 +28,4 @@ public class MainActivity extends Activity {
 			setContentView(R.layout.activity_main);
 		}
 	}
-	
-	public void onResume(){
-		super.onResume();
-		//himnokod
-		//Appl.appContext = this;
-	}
-
-	@Override
-	public boolean onCreateOptionsMenu(Menu menu) {
-		// Inflate the menu; this adds items to the action bar if it is present.
-		getMenuInflater().inflate(R.menu.main, menu);
-		return true;
-	}
-	
-	public void onPause(){
-		super.onPause();
-		//himnokod
-		//Appl.appContext=null;
-	}
-
 }
